Updated scenario

+++ src/main/scala/api/GenerationAPI.scala
@@ -1,6 +1,6 @@
 package api
 
-import java.io.PrintWriter
+import java.io.{File, PrintWriter}
 import java.nio.file.Files
 
 import cli.VerifyTA
@@ -13,69 +13,68 @@ import trace_analyzer.TraceAnalyzer
 
 object GenerationAPI {
 
-  def generateInitialization(scenarioModel: ScenarioModel)= {
+  def generateInitialization(scenarioModel: ScenarioModel) = {
     val synthesizer = new SynthesizerSimple(scenarioModel)
     synthesizer.synthesizeInitialization()
   }
 
-  def generatecoSimStep(scenarioModel: ScenarioModel)= {
+  def generatecoSimStep(scenarioModel: ScenarioModel) = {
     val synthesizer = new SynthesizerSimple(scenarioModel)
     synthesizer.synthesizeStep()
   }
 
   def generateAlgorithm(name: String, scenarioModel: ScenarioModel): MasterModel = {
-      val instantiationModel = generateInstantiationInstructions(scenarioModel).toList
-      val expandedInitModel = generateEnterInitInstructions(scenarioModel) ++ generateInitialization(scenarioModel) ++ generateExitInitInstructions(scenarioModel)
-      val cosimStepModel = generatecoSimStep(scenarioModel)
-      val terminateModel = generateTerminateInstructions(scenarioModel).toList
-      MasterModel(name, scenarioModel, instantiationModel, expandedInitModel.toList, cosimStepModel, terminateModel)
+    val instantiationModel = generateInstantiationInstructions(scenarioModel).toList
+    val expandedInitModel = generateEnterInitInstructions(scenarioModel) ++ generateInitialization(scenarioModel) ++ generateExitInitInstructions(scenarioModel)
+    val cosimStepModel = generatecoSimStep(scenarioModel)
+    val terminateModel = generateTerminateInstructions(scenarioModel).toList
+    MasterModel(name, scenarioModel, instantiationModel, expandedInitModel.toList, cosimStepModel, terminateModel)
   }
 }
 
 
 object VerificationAPI extends Logging {
-  def writeToTempFile(content: String) = {
+  private def writeToTempFile(content: String) = {
     val file = Files.createTempFile("uppaal_", ".xml").toFile
-    new PrintWriter(file) { write(content); close() }
+    new PrintWriter(file) {
+      write(content); close()
+    }
     file
   }
 
   def verifyAlgorithm(masterModel: MasterModel): Boolean = {
-    val encoding = new ModelEncoding(masterModel)
-    val result = ScenarioGenerator.generate(encoding)
-    val f = writeToTempFile(result)
-    assert(VerifyTA.checkEnvironment(), "UPPAAL v.4.1 is not in PATH - please install it and try again!")
+    val f = generateUppaalFile(masterModel)
+    checkUppaalVersion()
     val verificationResult = VerifyTA.verify(f)
     FileUtils.deleteQuietly(f)
 
-    if(verificationResult == 0) true
-    else {
-      if(verificationResult == 2)
-        throw SyntaxException("The verification in Uppaal failed most likely due to a syntax error in the UPPAAL model.")
-      else false
-    }
+    checkVerificationResult(verificationResult)
   }
 
-  def generateAndVerify(name: String, scenarioModel: ScenarioModel)= {
+  def generateAndVerify(name: String, scenarioModel: ScenarioModel) = {
     val masterModel = GenerationAPI.generateAlgorithm(name, scenarioModel)
     verifyAlgorithm(masterModel)
   }
 
-  def generateTrace(name: String, scenarioModel: ScenarioModel) = {
+  def generateTrace(name: String, scenarioModel: ScenarioModel):TraceResult = {
     val masterModel = GenerationAPI.generateAlgorithm(name, scenarioModel)
     generateTraceFromMasterModel(masterModel)
   }
 
-  def generateTraceFromMasterModel(masterModel: MasterModel) = {
+  private def generateUppaalFile(masterModel: MasterModel): File = {
     val encoding = new ModelEncoding(masterModel)
     val encodedUppaal = ScenarioGenerator.generate(encoding)
-    val f = writeToTempFile(encodedUppaal)
+    writeToTempFile(encodedUppaal)
+  }
+
+  def generateTraceFromMasterModel(masterModel: MasterModel): TraceResult = {
+    if(verifyAlgorithm(masterModel))
+      return TraceResult(null, false)
+
+    val f = generateUppaalFile(masterModel)
+    val encoding = new ModelEncoding(masterModel)
     val traceFile = Files.createTempFile("trace_", ".log").toFile
     val videoFile = Files.createTempFile("trace_", ".mp4").toFile
-
-    assert(VerifyTA.checkEnvironment(), "UPPAAL v.4.1 is not in PATH - please install it and try again!")
-    assert(VerifyTA.verify(f) != 0, "No trace can be generated - since the model is valid.")
-    assert(VerifyTA.verify(f) != 2, "No trace can be generated - the verification failed due to a syntax error.")
 
     VerifyTA.saveTraceToFile(f, traceFile)
     FileUtils.deleteQuietly(f)
@@ -86,10 +85,29 @@ object VerificationAPI extends Logging {
     }
     finally source.close()
     FileUtils.deleteQuietly(traceFile)
-    videoFile
+    TraceResult(videoFile, true)
+  }
+
+  def checkUppaalVersion() = {
+    if (!VerifyTA.checkEnvironment())
+      throw UppaalException("UPPAAL v.4.1 is not in PATH - please install it and try again!")
+  }
+
+  def checkVerificationResult(verificationResult: Int) = {
+    verificationResult match {
+      case 0 => true
+      case 2 => throw SyntaxException("The verification in Uppaal failed most likely due to a syntax error in the UPPAAL model.")
+      case _ => false
+    }
   }
 }
 
+final case class TraceResult(file: File, isGenerated: Boolean)
+
 final case class SyntaxException(private val message: String = "",
+                                 private val cause: Throwable = None.orNull)
+  extends Exception(message, cause)
+
+final case class UppaalException(private val message: String = "",
                                  private val cause: Throwable = None.orNull)
   extends Exception(message, cause)
